--- a/src/main/scala/com/gu/invoicing/invoice/Model.scala
+++ b/src/main/scala/com/gu/invoicing/invoice/Model.scala
@@ -83,7 +83,7 @@ object Model extends OptionPickler {
     date: LocalDate,
     paymentMethod: PaymentMethod,
     price: Double,
-    pdfFileUrl: String, /* invoices/{accountId}/{fileId} */
+    pdfPath: String, /* invoices/{accountId}/{fileId} */
     invoiceId: String
   )
 
@@ -98,7 +98,7 @@ object Model extends OptionPickler {
     ): InvoiceWithPayment = {
       // PDF invoice files are in reverse chronological order meaning most recent version is first
       // https://www.zuora.com/developer/api-reference/#operation/GET_InvoiceFiles
-      val pdfFileUrl =
+      val pdfPath =
         invoice
           .invoiceFiles
           .headOption
@@ -118,7 +118,7 @@ object Model extends OptionPickler {
         date = invoice.invoiceDate,
         paymentMethod = paymentMethod,
         price = invoice.amount.toDouble,
-        pdfFileUrl = pdfFileUrl,
+        pdfPath = pdfPath,
         invoiceId = invoice.id
       )
     }
@@ -132,7 +132,7 @@ object Model extends OptionPickler {
     invoiceId: String,
     subscriptionName: String,
     date: LocalDate,
-    pdfFileUrl: String,
+    pdfPath: String,
     price: Double,
     paymentMethod: String,
     last4: Option[String] = None // for card and direct debit
@@ -144,7 +144,7 @@ object Model extends OptionPickler {
         invoiceId = invoiceWithPayment.invoiceId,
         subscriptionName = invoiceWithPayment.subscriptionName,
         date = invoiceWithPayment.date,
-        pdfFileUrl = invoiceWithPayment.pdfFileUrl,
+        pdfPath = invoiceWithPayment.pdfPath,
         price = invoiceWithPayment.price,
         paymentMethod = invoiceWithPayment.paymentMethod.Type
       )