--- a/src/main/scala/com/gu/invoicing/invoice/README.md
+++ b/src/main/scala/com/gu/invoicing/invoice/README.md
@@ -25,7 +25,7 @@ Authorization: ********
             "invoiceId": "1234567899qwertyuio",
             "subscriptionName": "A-S0000000",
             "date": "2020-08-05",
-            "pdfFileUrl": "invoices/anAccountId/afileId",
+            "pdfPath": "invoices/anAccountId/afileId",
             "price": 813.48,
             "paymentMethod": "BankTransfer",
             "last4": "9911"