## What does this change?

https://trello.com/c/ggqXe1rg/1492-goal-expose-invoices-for-the-billing-tab

Return 

```
{
    "invoices": [
        {
            "invoiceId": "1234567899qwertyuio",
            "subscriptionName": "A-S0000000",
            "date": "2020-08-05",
            "pdfPath": "invoices/afileId",
            "price": 813.48,
            "paymentMethod": "BankTransfer",
            "last4": "9911"
        }
    ]
}
```

when server-side of manage-frontend requests `GET /invoices`, which will be rendered as something like

![image](https://user-images.githubusercontent.com/13835317/89781931-2c724c00-db0c-11ea-9a48-84a1f9f152d2.png)

### Key implementation details

- The field `"pdfPath":  "invoices/afileId"` corresponds to `fileId` of Zuora [GET_InvoiceFiles](https://www.zuora.com/developer/api-reference/#operation/GET_InvoiceFiles) which will be used to download the PDF invoice file (implemented in separate PR). 
- MMA will need to send an x-identity-id + API key + [AWS signature](https://github.com/guardian/manage-frontend/pull/465).
- Main business logic is in `Program.scala`
- Implementation uses Scala 2.13.3 [`-Xasync`](https://github.com/scala/scala-async) compiler flag to enable the following direct style of programming

```
val invoicesF              = async(getInvoices(accountId))
val paymentsF              = async(getPayments(accountId))
val paymentMethodsF        = async(getPaymentMethods(accountId))
val positiveInvoices       = await(invoicesF)
val payments               = await(paymentsF)
val paymentMethods         = await(paymentMethodsF)
val _                      = allInvoicesShouldHaveASingleSubscription(positiveInvoices)
val invoicesWithPayment    = joinInvoicesWithPayment(positiveInvoices, payments, paymentMethods)
val mmaInvoicesWithPayment = transformToMmaExpectedFormat(invoicesWithPayment)
```

instead of equivalent monadic for-comprehension 

```
val invoicesF              = Future(getInvoices(accountId)))
val paymentsF              = Future(getPayments(accountId)))
val paymentMethodsF        = Future(getPaymentMethods(accountId)))
for {
  positiveInvoices <- invoicesF
  payments         <- paymentsF
  paymentMethods   <- paymentMethodsF
} yield ...
```
- Performance breakdown is approx. as follows
  - cold =  900 ms Lambda warmup + 500 ms Zuora auth + 1600 ms business logic
  - warm = 1500 ms total
- Lambda package size at approx 7 MB
- Approach of fail "deny + fail fast + alarm + iterate" to root out edge cases before thinking about generalisation

### Limitations

- It does not handle invoices with multiple subscriptions

## How to test

#### With CLI 

1. Get invoicing-api+uat@guardian.co.uk Zuora OAuth client credentials
1. Follow docs in `Cli.scala` and create `Stage` and `Config` environmental variables
1. Performance can be estimated with something like 
    ```
    time { Await.result(program(InvoicesInput("someAccountId")), Inf) }
    ```

#### With Lambda

1. Get fresh Janus credentials
1. `Program.scala` contains main business logic 
1. `deployAwsLambda` will upload modified lambda package
1. In AWS Lambda console create test event with reflects `GET /invoices/{accountId}`
    ```
    {
      "pathParameters": {
        "accountId": "1234567qwerty"
      }
    }
    ```
1. Tail logs 
    ```
    export AWS_DEFAULT_REGION=eu-west-1 && awslogs get --profile membership /aws/lambda/invoicing-api-invoices-CODE --watch
    ```
   
#### With API Gateway (Postman)

1. HTTP request will have to be signed with AWS signature
1. Get fresh janus credentials
1. Paste them in `Postman | Auth | AWS Signature`
1. Don't forget also the API key `x-api-key`


## How can we measure success?

This endpoint should not be returning non-200 responses because the client is server side of manage-frontend which should always have valid credentials, and send request for invoices only for users that have paid for a product which MMA can check via members-data-api. This means failure likely represents misconfiguration, or missing implementation, so should be fixed ASAP.

## Have we considered potential risks?

- infosec has been consulted
- additional AWS signature is required for all requests on top of API key
- signed in user should not be able to request another user's invoices
- alarm on failure