@mario-galic 2020-08-12
@twrichards Doesn't manage anyway have to call mem-data-api as billing is loaded so we just piggyback on that?